chore(cli,emit): split oversized modules below §19 cap
Split `driver/resolution.rs` (4109 lines) into three focused modules:
- `resolution.rs` (1469 lines): infrastructure, type packages, path utils
- `resolution_specifier_scanning.rs` (1248 lines): specifier collection/scanning
- `resolution_package_lookup.rs` (1446 lines): npm/package resolution logic

Each sub-module imports parent items via `super::` and exposes its public
surface via `pub(super)` or module-path prefixes in the parent. No behavior
change; pure structural refactor.

Also split `transforms/module_commonjs.rs` (2016 lines) by extracting the
inline `mod tests` block into `module_commonjs_tests.rs` (1971 lines
remaining), keeping both files under the 2000-line cap.

Closes #10709.
